Changed How the Portal Works
When entering a portal, the player exists the adjacent direction they came in (going in left of the portal will exit right of the destination portal).
Otherwise, the next available will be the destination.
This makes it possible for the player to land into enemies and removes the exploit of staying in the destination portal.